pyproject.toml
[tool.poetry]
name = "campgames_bot"
version = "0.1.0"
description = ""
authors = ["Vlad-Vargan <blach.smith.2000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["ext"], version = "^20.3"}
python-dotenv = "^0.19.2"
gspread = "^5.1.1"
oauth2client = "^4.1.3"
loguru = "^0.5.3"
pydantic = "^1.10.8"
sqlalchemy = "^2.0.15"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
black = {version = "^21.12b0", allow-prereleases = true}

[tool.poetry.group.dev.dependencies]
sqlalchemy = {extras = ["mypy"], version = "^2.0.15"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
